+++ pandas/tests/frame/test_api.py
@@ -6,11 +6,11 @@ import numpy as np
 import pytest
 
 from pandas.compat import PY37
+from pandas.util._test_decorators import async_mark
 
 import pandas as pd
 from pandas import Categorical, DataFrame, Series, compat, date_range, timedelta_range
 import pandas._testing as tm
-from pandas.conftest import async_mark
 
 
 class TestDataFrameMisc:
