+++ pandas/tests/indexes/test_base.py
@@ -12,6 +12,7 @@ import pandas._config.config as cf
 
 from pandas._libs.tslib import Timestamp
 from pandas.compat.numpy import np_datetime64_compat
+from pandas.util._test_decorators import async_mark
 
 from pandas.core.dtypes.common import is_unsigned_integer_dtype
 from pandas.core.dtypes.generic import ABCIndex
@@ -33,7 +34,6 @@ from pandas import (
     period_range,
 )
 import pandas._testing as tm
-from pandas.conftest import async_mark
 from pandas.core.algorithms import safe_sort
 from pandas.core.indexes.api import (
     Index,
